+++ flake.nix
@@ -1,19 +1,22 @@
 {
   description = "Gluten Nix Flake";
 
+  inputs.nix-filter.url = "github:numtide/nix-filter";
   inputs.flake-utils.url = "github:numtide/flake-utils";
   inputs.nixpkgs.inputs.flake-utils.follows = "flake-utils";
   inputs.nixpkgs.url = "github:anmonteiro/nix-overlays";
 
-  outputs = { self, nixpkgs, flake-utils }:
+  outputs = { self, nixpkgs, flake-utils, nix-filter }:
     flake-utils.lib.eachDefaultSystem (system:
       let
         pkgs = nixpkgs.legacyPackages."${system}".extend (self: super: {
-          ocamlPackages = super.ocaml-ng.ocamlPackages_4_14;
+          ocamlPackages = super.ocaml-ng.ocamlPackages_5_00;
         });
+        packages = pkgs.callPackage ./nix { nix-filter = nix-filter.lib; };
+
       in
-      rec {
-        packages = (pkgs.callPackage ./nix { }) // { default = packages.gluten; };
-        devShells.default = pkgs.callPackage ./shell.nix { };
+      {
+        packages = packages // { default = packages.gluten; };
+        devShells.default = pkgs.callPackage ./shell.nix { inherit packages; };
       });
 }