+++ nix/ci/test.nix
@@ -5,16 +5,23 @@ let
   src = fetchGit {
     url = with lock.nodes.nixpkgs.locked;"https://github.com/${owner}/${repo}";
     inherit (lock.nodes.nixpkgs.locked) rev;
-    # inherit (lock.nodes.nixpkgs.original) ref;
+    allRefs = true;
   };
   pkgs = import src {
-    overlays = [
-      (import src)
+    extraOverlays = [
       (self: super: {
         ocamlPackages = super.ocaml-ng."ocamlPackages_${ocamlVersion}";
       })
     ];
   };
+  nix-filter-src = fetchGit {
+    url = with lock.nodes.nix-filter.locked; "https://github.com/${owner}/${repo}";
+    inherit (lock.nodes.nix-filter.locked) rev;
+    # inherit (lock.nodes.nixpkgs.original) ref;
+    allRefs = true;
+  };
+  nix-filter = import "${nix-filter-src}";
 
 in
-pkgs.callPackages ./.. { doCheck = true; }
+
+pkgs.callPackage ./.. { doCheck = true; inherit nix-filter; }